+++ tests/testthat/test-input.R
@@ -2,60 +2,58 @@ context("input")
 
 out <- c("``` r", "1:5", "#> [1] 1 2 3 4 5", "```")
 
-exp_msg <- switch(as.character(clipboard_available()),
-                  `TRUE` = "Rendered reprex is on the clipboard.",
-                  "Unable to put result on the clipboard")
+exp_msg <- switch(
+  as.character(clipboard_available()),
+  `TRUE` = "Rendered reprex is on the clipboard.",
+  "Unable to put result on the clipboard"
+)
 
 test_that("reprex: clipboard input works", {
   skip_if_no_clipboard()
   clipr::write_clip("1:5")
-  expect_message(ret <- reprex(show = FALSE), exp_msg)
-  expect_identical(ret, out)
+  expect_match(reprex(render = FALSE), "^1:5$", all = FALSE)
 })
 
 test_that("reprex: expression input works", {
-  msg <- capture_messages(ret <- reprex(1:5, show = FALSE))
-  expect_identical(ret, out)
-  expect_match(msg, exp_msg, all = FALSE)
+  expect_match(reprex(1:5, render = FALSE), "^1:5$", all = FALSE)
 })
 
 test_that("reprex: character input works", {
-  msg <- capture_messages(ret <- reprex(input = "1:5\n", show = FALSE))
-  expect_identical(ret, out)
-  expect_match(msg, exp_msg, all = FALSE)
+  expect_match(reprex(input = "1:5\n", render = FALSE), "^1:5$", all = FALSE)
 })
 
 test_that("reprex: file input works", {
   on.exit(file.remove("foo.R"))
   write("1:5", "foo.R")
-  msg <- capture_messages(ret <- reprex(input = "foo.R", show = FALSE))
-  expect_identical(ret, out)
-  expect_match(msg, exp_msg, all = FALSE)
+  expect_match(reprex(input = "foo.R", render = FALSE), "^1:5$", all = FALSE)
 })
 
 test_that("reprex: file input in a subdirectory works", {
   on.exit(unlink("foo", recursive = TRUE))
   dir.create("foo")
   write("1:5", file.path("foo", "foo.R"))
-  msg <- capture_messages(ret <- reprex(input = file.path("foo", "foo.R"),
-                                        show = FALSE))
-  expect_identical(ret, out)
-  expect_match(msg, exp_msg, all = FALSE)
+  expect_match(
+    reprex(input = file.path("foo", "foo.R"), render = FALSE),
+    "^1:5$",
+    all = FALSE
+  )
 })
 
-test_that("Circular use is detected before render", {
+test_that("Circular use is detected before source file written", {
   ret <- reprex(y <- 2, venue = "gh", show = FALSE)
-  expect_error(reprex(input = ret, show = FALSE), "isn't valid R code")
+  expect_error(reprex(input = ret, render = FALSE), "isn't valid R code")
   ret <- reprex(y <- 2, venue = "so", show = FALSE)
-  expect_error(reprex(input = ret, show = FALSE), "isn't valid R code")
+  expect_error(reprex(input = ret, render = FALSE), "isn't valid R code")
 })
 
 test_that("Leading prompts are removed", {
   input <- c("x <- 1:3", "median(x)")
-  res <- reprex(input = input, show = FALSE)
+  res <- reprex(input = input, render = FALSE)
   input2 <- paste0(getOption("prompt"), input)
-  expect_message(res2 <- reprex(input = input2, show = FALSE),
-                 "Removing leading prompts")
+  expect_message(
+    res2 <- reprex(input = input2, render = FALSE),
+    "Removing leading prompts"
+  )
   expect_identical(res, res2)
 })